refactor: move card action type
